Move ReportBatch
No changes to the type, but now all of the "data" types are in
report.go.
